+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New:
 ----------------------
 - We added a functionality to create ingredients replacements data
+- Footer is now sticky formula view
+